refactor: storybook layout
