+++ src/examples/GameDev/HexGrid/CustomHex.ts
