+++ src/examples/GameDev/HexGrid/HexGrid.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { HexGridScene } from "./HexGridScene";
 
 const meta = {
-  title: "Examples/ReactThreeJs/HexGrid",
+  title: "Examples/Game Dev/Hex Grid",
   component: HexGridScene,
   parameters: {
     layout: "fullscreen",
