+++ src/examples/GameDev/HexGrid/Models/HexTile.tsx
