+++ src/examples/GameDev/HexGrid/Models/HexTileGrass/HexTileGrass.tsx
