+++ src/examples/GameDev/HexGrid/Models/HexTileGrass/index.ts
