+++ src/examples/GameDev/HexGrid/Models/Overlay/Highlight.tsx
