+++ src/examples/GameDev/HexGrid/Models/Overlay/Outline.tsx
