+++ src/examples/GameDev/HexGrid/Models/Overlay/Sphere.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Scene } from "../Scene";
-import { Player as PlayerModel } from "./Player";
+import { OverlaySphere } from "./Sphere";
+import { Scene } from "..";
 
 const meta = {
-  title: "Examples/ReactThreeJs/HexGrid/Player",
+  title: "Examples/Game Dev/Hex Grid/Overlays",
   decorators: [
     (Story) => (
       <Scene>
@@ -12,13 +12,13 @@ const meta = {
       </Scene>
     ),
   ],
-  component: PlayerModel,
+  component: OverlaySphere,
   parameters: {
     layout: "fullscreen",
   },
-} satisfies Meta<typeof PlayerModel>;
+} satisfies Meta<typeof OverlaySphere>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Player: Story = {};
+export const Sphere: Story = {};
