+++ src/examples/GameDev/HexGrid/Models/Overlay/Sphere.tsx
