+++ src/examples/GameDev/HexGrid/Models/Overlay/Text.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { OverlayText } from "./Text";
-import { Scene } from "../";
+import { Scene } from "..";
 import { TILE_COLORS } from "../../constants";
 
 const meta = {
-  title: "Examples/ReactThreeJs/HexGrid/Overlays",
+  title: "Examples/Game Dev/Hex Grid/Overlays",
   decorators: [
     (Story) => (
       <Scene>
