+++ src/examples/GameDev/HexGrid/Models/Overlay/index.ts
