+++ src/examples/GameDev/HexGrid/Models/Player/Player.tsx
