+++ src/examples/GameDev/HexGrid/Models/Player/PlayerScene.tsx
