+++ src/examples/GameDev/HexGrid/Models/Player/index.ts
