+++ src/examples/GameDev/HexGrid/Notes.mdx
@@ -1,4 +1,15 @@
+import { Meta } from "@storybook/blocks";
+import LinkTo from "@storybook/addon-links/react";
+
+<Meta
+  name="Examples/Game Dev/Hex Grid/Notes"
+  title="Examples/Game Dev/Hex Grid/Notes"
+/>
+
 # HexGrid Working Notes
+
+<time dateTime="2024-02-08">Feb 8, 2024</time>
+
 
 - [Honeycomb Hex Grid Library][2]
 - [Hex Geometry][3]
