+++ src/examples/GameDev/HexGrid/_tests_/CustomHex.test.ts
