+++ src/examples/GameDev/HexGrid/_tests_/utils/calculateRotation.test.ts
