+++ src/examples/GameDev/HexGrid/constants.ts
