+++ src/examples/GameDev/HexGrid/honeycomb.d.ts
