+++ src/examples/GameDev/HexGrid/store.ts
