+++ src/examples/GameDev/HexGrid/types.ts
