+++ src/examples/GameDev/HexGrid/utils/calculateDirection.ts
