+++ src/examples/GameDev/HexGrid/utils/calculateRotation.ts
