+++ src/examples/GameDev/HexGrid/utils/calculateTileRotation.ts
