+++ src/examples/GameDev/HexGrid/utils/getRandomItemIndex.ts
