+++ src/examples/GameDev/HexGrid/utils/getTileOverlayColor.ts
