+++ src/examples/Three/Avatar/AvatarModel.tsx
