+++ src/examples/Three/Avatar/AvatarScene.tsx
