+++ src/notes/3d-rendering/ThreeJsWithReact.mdx
@@ -1,6 +1,6 @@
 import { Source, Meta, Canvas } from "@storybook/blocks";
 
-import * as AvatarStory from "../../examples/ReactThreeJs/Avatar/Avatar.stories";
+import * as AvatarStory from "../../examples/Three/Avatar/Avatar.stories";
 
 import UEMannequins from "../../assets/screenshots/UEMannequins.png";
 import MixamoCharacterRigging from "../../assets/screenshots/MixamoCharacterRigging.png";
